+++ tests/schema/test_response_schema.py
@@ -4,10 +4,34 @@ from typing import Type
 import pytest
 from drf_yasg import openapi
 
+import winter
 from winter.controller import get_component
 from winter.routing.routing import get_route
+from winter.schema.generation import build_response_schema
 from winter.schema.generation import build_responses_schemas
 from ..controllers import ControllerWithExceptions
+
+
+@pytest.mark.parametrize(('return_type', 'expected_response'), (
+        (str, openapi.Schema(type=openapi.TYPE_STRING)),
+        (None, openapi.Response(description='')),
+        (object, openapi.Response(description='')),
+))
+def test_build_response_schema(return_type, expected_response):
+    @winter.route('')
+    class SimpleController:
+
+        @winter.route_get('/simple/')
+        def simple_method(self) -> return_type:
+            return None
+
+    route = get_route(SimpleController.simple_method)
+
+    # Act
+    schema = build_response_schema(route)
+
+    # Assert
+    assert isinstance(schema, openapi.SwaggerDict)
 
 
 @pytest.mark.parametrize('controller_class, method_name, expected_responses', [
